sr-iov: add logLevel to SriovOperatorConfig
Starting from OCP 4.16 GA, the SriovOperatorConfig is no longer
automatically created by the Operator upon install. Instead it must be
created by cluster administrators.

Option logLevel should be added to SriovOperatorConfig.yaml since it is
the default value but more importantly because it is documented as an
option users should set when creating the SriovOperatorConfig CR.

* https://docs.openshift.com/container-platform/4.16/networking/hardware_networks/configuring-sriov-operator.html