trying to set some order for optional components
Trying to have more control on components deployment.
For optional odf and lso operators are installed at
once. Then, it goes the lso volumes creation and
it does not continue until everything is oka. After
that, we create the odf cluster with a final validation.

If the validation went oka, we are in the wave 0,
and we continue with the other components.