telco-hub: Reduce ArgoCD ClusterRole with minimal privilege
The current cluster-admin ClusterRole assigned to ArgoCD in Hub clusters grants excessive permissions that are not necessary for its intended functionality. By adopting a minimal privilege ClusterRole, we align with the Principle of Least Privilege, reducing the attack surface and limiting access to only the required resources. This minimizes the potential for accidental misuse, privilege escalation, or unauthorized access in the event of a compromise.

Signed-off-by: Leonardo Ochoa-Aday <lochoa@redhat.com>